feat: move code to a 'new' folder
