+++ multiauth/new/entities/__init__.py
@@ -1,0 +1,1 @@
+"""Types and entities for multiauth."""
