+++ multiauth/new/entities/apikey.py
@@ -1,0 +1,5 @@
+from multiauth.new.entities.main import Credentials
+
+
+class APIKeyCredentials(Credentials):
+    pass
