+++ multiauth/new/entities/basic.py
@@ -1,0 +1,6 @@
+from multiauth.new.entities.main import Credentials
+
+
+class BasicCredentials(Credentials):
+    username: str
+    password: str
