+++ multiauth/new/entities/digest.py
@@ -1,0 +1,6 @@
+from multiauth.new.entities.main import Credentials
+
+
+class DigestCredentials(Credentials):
+    username: str
+    password: str
