+++ multiauth/new/entities/rest.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class RESTCredentials(BaseModel):
+    pass
