+++ multiauth/providers/apikey.py
@@ -3,9 +3,9 @@
 from typing import cast
 
 from multiauth.entities.errors import AuthenticationError
+from multiauth.entities.http import HTTPLocation
 from multiauth.entities.main import AuthResponse, AuthTech, Token
 from multiauth.entities.providers.apikey import AuthConfigApiKey
-from multiauth.entities.providers.http import HTTPLocation
 from multiauth.manager import User
