+++ multiauth/providers/rest.py
@@ -7,9 +7,8 @@ import jwt
 import requests
 
 from multiauth.entities.errors import AuthenticationError
-from multiauth.entities.http import HTTPHeaders
+from multiauth.entities.http import HTTPHeaders, HTTPLocation
 from multiauth.entities.main import AuthResponse, AuthTech, Token
-from multiauth.entities.providers.http import HTTPLocation
 from multiauth.entities.providers.rest import AuthConfigRest, CredentialsEncoding
 from multiauth.helpers import extract_token
 from multiauth.manager import User
